controller to list created
