+++ src/routes/cursos.route.js
@@ -2,6 +2,7 @@ const { Router, query } = require('express') //
 const Curso = require('../models/Curso')
 
 const { auth } = require('../middleware/auth')
+const CursoController = require('../controllers/CursoController')
 
 const cursoRoutes = new Router()
 
@@ -34,33 +35,7 @@ cursoRoutes.post('/', auth,  async (req, res) => {
 })
 
 
-cursoRoutes.get('/', auth,  async (req, res) => {
-    try {
-        let params = {}
-
-        // SE for passado uma paramero QUERY chamado "nome" na requisição, então
-        // esse parametro "nome" é adicionado dentro da variavel params
-        if (req.query.nome) {
-            // o ...params, cria uma cópia do params com os chaves e valores já existentes
-            params = { ...params, nome: req.query.nome }
-        }
-
-        if (req.query.duracao_horas) {
-            // o ...params, cria uma cópia do params com os chaves e valores já existentes
-            params = { ...params, duracao_horas: req.query.duracao_horas }
-        }
-
-        const cursos = await Curso.findAll({
-            where: params
-        })
-
-        res.json(cursos)
-    } catch (error) {
-        console.log(error.message)
-        res.status(500).json({ error: 'Não possível listar todos os cursos' })
-    }
-})
-
+cursoRoutes.get('/', auth, CursoController.listar)
 
 cursoRoutes.delete('/:id', auth, (req, res) => {
     const { id } = req.params
